+++ components/TextControls/tests/snapshot/MDCOutlinedTextAreaSnapshotTests.m
@@ -12,14 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import "MDCTextControl.h"
-#import "MDCTextControlAssistiveLabelView.h"
-#import "MDCTextControlAssistiveLabelViewLayout.h"
-#import "MDCTextControlColorViewModel.h"
-#import "MDCTextControlGradientManager.h"
-#import "MDCTextControlHorizontalPositioning.h"
-#import "MDCTextControlHorizontalPositioningReference.h"
-#import "MDCTextControlLabelAnimation.h"
-#import "MDCTextControlLabelPosition.h"
-#import "MDCTextControlPlaceholderSupport.h"
-#import "MDCTextControlVerticalPositioningReference.h"
+#import "MDCBaseTextAreaSnapshotTests.h"
+
+#import <UIKit/UIKit.h>
+
+#import "MDCOutlinedTextArea.h"
+
+@interface MDCOutlinedTextAreaSnapshotTests : MDCBaseTextAreaSnapshotTests
+@end
+
+@implementation MDCOutlinedTextAreaSnapshotTests
+
+- (MDCBaseTextArea *)createTextAreaWithFrame:(CGRect)frame {
+  return (MDCBaseTextArea *)[[MDCOutlinedTextArea alloc] initWithFrame:frame];
+}
+
+@end