+++ components/private/TextControlsPrivate/src/TextFields/MDCBaseTextFieldLayout.h
@@ -17,16 +17,16 @@
 
 @interface MDCBaseTextFieldLayout : NSObject
 
-@property(nonatomic, assign) BOOL leftViewHidden;
-@property(nonatomic, assign) BOOL rightViewHidden;
+@property(nonatomic, assign) BOOL displaysLeadingView;
+@property(nonatomic, assign) BOOL displaysTrailingView;
 
 @property(nonatomic, assign) CGRect clearButtonFrame;
 @property(nonatomic, assign) CGRect labelFrameFloating;
 @property(nonatomic, assign) CGRect labelFrameNormal;
 @property(nonatomic, assign) CGRect textRectNormal;
 @property(nonatomic, assign) CGRect textRectFloating;
-@property(nonatomic, assign) CGRect leftViewFrame;
-@property(nonatomic, assign) CGRect rightViewFrame;
+@property(nonatomic, assign) CGRect leadingViewFrame;
+@property(nonatomic, assign) CGRect trailingViewFrame;
 @property(nonatomic, assign) CGRect assistiveLabelViewFrame;
 @property(nonatomic, strong, nonnull)
     MDCTextControlAssistiveLabelViewLayout *assistiveLabelViewLayout;
@@ -55,10 +55,10 @@
                        labelPosition:(MDCTextControlLabelPosition)labelPosition
                        labelBehavior:(MDCTextControlLabelBehavior)labelBehavior
                    sideViewAlignment:(MDCTextControlTextFieldSideViewAlignment)sideViewAlignment
-                            leftView:(nullable UIView *)leftView
-                        leftViewMode:(UITextFieldViewMode)leftViewMode
-                           rightView:(nullable UIView *)rightView
-                       rightViewMode:(UITextFieldViewMode)rightViewMode
+                         leadingView:(nullable UIView *)leadingView
+                     leadingViewMode:(UITextFieldViewMode)leadingViewMode
+                        trailingView:(nullable UIView *)trailingView
+                    trailingViewMode:(UITextFieldViewMode)trailingViewMode
                clearButtonSideLength:(CGFloat)clearButtonSideLength
                      clearButtonMode:(UITextFieldViewMode)clearButtonMode
                leadingAssistiveLabel:(nonnull UILabel *)leftAssistiveLabel